Add files via upload
How to add go to top button using HTML, CSS, and JavaScript as we have used in our website Solutiongigs, how to add simple and easy scroll to top button if we are using WordPress or Coding without any complications. There are many methods to add go to top button in WordPress basically, and for coding, just a few lines of code are shown below. for more please visit